have one less thread switch for re]play
