+++ sentry-android-replay/src/main/java/io/sentry/android/replay/util/Executors.kt
@@ -50,6 +50,11 @@ internal fun ExecutorService.submitSafely(
     taskName: String,
     task: Runnable
 ): Future<*>? {
+    if (Thread.currentThread().name.startsWith("SentryReplayIntegration")) {
+        // we're already on the worker thread, no need to submit
+        task.run()
+        return null
+    }
     return try {
         submit {
             try {
@@ -73,7 +78,7 @@ internal fun ScheduledExecutorService.scheduleAtFixedRateSafely(
     task: Runnable
 ): ScheduledFuture<*>? {
     return try {
-        scheduleAtFixedRate({
+        scheduleWithFixedDelay({
             try {
                 task.run()
             } catch (e: Throwable) {
